+++ EnhanceQoL/Locales/ptBR.lua
@@ -34,6 +34,7 @@ L["MissingEnchant"] = "Encantamento"
 L["hideHitIndicatorPlayer"] = "隱藏角色頭上的浮動戰鬥文字（傷害和治療）"
 L["hideHitIndicatorPet"] = "隱藏寵物頭上的浮動戰鬥文字（傷害和治療）"
 L["UnitFrame"] = "單位框架"
+L["SellJunkIgnoredBag"] = "Você desativou a venda de itens inúteis em %d bolsas.\nIsso pode impedir a venda automática de todos os itens inúteis."
 
 L["deleteItemFillDialog"] = 'Adicionar "' .. DELETE_ITEM_CONFIRM_STRING .. '" ao "Popup de confirmação de exclusão"'
 L["confirmPatronOrderDialog"] = "Confirma automaticamente o uso de materiais próprios em encomendas de " .. PROFESSIONS_CRAFTER_ORDER_TAB_NPC