+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Added:
 
+- When you check "Automatically sell all junk items" it will check if any of your bags is set to "Ignore sell junk" and reports that.
+
 ### New Options
 
 - Automatically confirm to use own materials on Patron crafting orders